LLM_server/requirement.txt
fastapi 
uvicorn 
sentence-transformers 
transformers 
torch 
tf-keras 
tensorflow
fastapi-cors
groq
nltk
google-generativeai